bugfix: 修复删除逻辑修改之后的单测逻辑

+++ apps/log_search/handlers/search/favorite_handlers.py
@@ -49,6 +49,7 @@ from apps.log_search.exceptions import (
     FavoriteNotExistException,
     FavoriteVisibleTypeNotAllowedModifyException,
     FavoriteAlreadyExistException,
+    FavoriteNotAllowedDeleteException,
 )
 from apps.log_search.models import Favorite, FavoriteGroup, FavoriteGroupCustomOrder, LogIndexSet
 
@@ -215,6 +216,9 @@ class FavoriteHandler(object):
             )
 
     def delete(self):
+        # 只有收藏的创建者才可以删除
+        if self.data.created_by != self.username:
+            raise FavoriteNotAllowedDeleteException()
         self.data.delete()
 
     @staticmethod
